Moved SoupHandler to a new file.
